pyproject.toml
[tool.poetry]
name = "mellow_strategy_sdk"
version = "0.1"
description = "Framework for creating new Uniswap V3 strategies"
authors = ["Your Name <you@example.com>"]
license = "MIT License"

packages = [
    {include = "strategy"},
    {include = 'examples'},
    {include = 'tests_unittests'},
    {include = 'utilities'}
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.0"
pandas = "^1.3.5"
plotly = "^5.5.0"
polars = "^0.13.2"
structlog = "^21.5.0"
pyarrow = "^7.0.0"
PyYAML = "^6.0"
SQLAlchemy = "^1.4.31"
parameterized = "^0.8.1"
sphinx-rtd-theme = "^1.0.0"
python-binance = "^1.0.15"
autodoc = "^0.5.0"
sphinx-autodocgen = "^1.3"
sphinxcontrib-napoleon = "^0.7"
update = "^0.0.1"
myst-parser = "^0.17.0"
sphinx-autodoc-typehints = "^1.17.0"
poetry = "^1.1.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
